Fix e2e smoke and CLI latency flush
