+++ package.json
@@ -105,8 +105,9 @@
     "test:unit": "vitest run --exclude test/package-exports.test.ts",
     "test:package-exports": "pnpm run build && VUE_TUI_REQUIRE_DIST_EXPORTS=1 vitest run test/package-exports.test.ts",
     "typecheck": "tsc --noEmit",
-    "e2e": "playwright test",
+    "e2e": "playwright test --project=smoke",
     "e2e:smoke": "playwright test --project=smoke",
+    "e2e:all": "playwright test",
     "bench:vfor": "tsx scripts/bench-vfor.ts",
     "bench:dom-renderer": "tsx scripts/bench-dom-renderer.ts",
     "bench:phase2": "tsx scripts/bench-phase2.ts",
