+++ src/cli.ts
@@ -14,7 +14,7 @@ export type { CreateTerminalAppOptions, TerminalApp } from "./create-terminal-ap
 export { createTerminalApp } from "./create-terminal-app.js";
 export type { CliEventManager } from "./events/manager/cli-event-manager.js";
 export { createCliEventManager } from "./events/manager/cli-event-manager.js";
-export { getCliLatencyProfiler } from "./observability/cli-latency.js";
+export { getCliLatencyProfiler } from "./observability/cli-latency-node.js";
 export { createOsc52ClipboardProvider, type Osc52ClipboardOptions } from "./runtime/index.js";
 export type { StdoutRendererMetrics } from "./renderer/cli/stdout-metrics.js";
 export { getStdoutRendererMetrics } from "./renderer/cli/stdout-metrics.js";
